+++ buildSrc/src/main/kotlin/tech/thdev/gradle/dependencies/Gradle.kt
@@ -1,6 +1,0 @@
-package tech.thdev.gradle.dependencies
-
-object Gradle {
-    const val android = "com.android.tools.build:gradle:7.0.4"
-    const val kotlin = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Dependency.Kotlin.version}"
-}