+++ flow-call-adapter-factory/src/main/java/tech/thdev/network/flowcalladapterfactory/internal/InternalUtil.kt
@@ -1,12 +1,13 @@
 package tech.thdev.network.flowcalladapterfactory.internal
 
-import kotlin.coroutines.resumeWithException
-import kotlinx.coroutines.CancellableContinuation
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import kotlin.coroutines.*
+import kotlinx.coroutines.*
+import retrofit2.*
 
-internal fun Call<*>.registerOnCancellation(continuation: CancellableContinuation<*>) {
+internal fun Call<*>.registerOnCancellation(
+    continuation: CancellableContinuation<*>,
+) {
+
     continuation.invokeOnCancellation {
         try {
             cancel()
@@ -16,14 +17,19 @@ internal fun Call<*>.registerOnCancellation(continuation: CancellableContinuatio
     }
 }
 
-internal fun <T> Call<T>.registerCallback(continuation: CancellableContinuation<*>, success: (response: Response<T>) -> Unit) {
-    enqueue(object : Callback<T> {
-        override fun onResponse(call: Call<T>, response: Response<T>) {
-            success(response)
-        }
+internal fun <T> Call<T>.registerCallback(
+    continuation: CancellableContinuation<*>,
+    success: (response: Response<T>) -> Unit,
+) {
+    enqueue(
+        object : Callback<T> {
+            override fun onResponse(call: Call<T>, response: Response<T>) {
+                success(response)
+            }
 
-        override fun onFailure(call: Call<T>, t: Throwable) {
-            continuation.resumeWithException(t)
+            override fun onFailure(call: Call<T>, t: Throwable) {
+                continuation.resumeWithException(t)
+            }
         }
-    })
+    )
 }