+++ flow-call-adapter-factory/src/main/java/tech/thdev/network/flowcalladapterfactory/internal/ResponseCallAdapter.kt
@@ -1,14 +1,9 @@
 package tech.thdev.network.flowcalladapterfactory.internal
 
-import java.lang.reflect.Type
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.suspendCancellableCoroutine
-import retrofit2.Call
-import retrofit2.CallAdapter
-import retrofit2.HttpException
-import retrofit2.Response
+import java.lang.reflect.*
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.*
+import retrofit2.*
 
 internal class ResponseCallAdapter<T>(
     private val responseType: Type
@@ -17,20 +12,21 @@ internal class ResponseCallAdapter<T>(
     override fun responseType(): Type =
         responseType
 
-    @OptIn(ExperimentalCoroutinesApi::class)
     override fun adapt(call: Call<T>): Flow<Response<T>> = flow {
-        emit(suspendCancellableCoroutine { continuation ->
-            call.registerCallback(continuation) { response ->
-                continuation.resumeWith(kotlin.runCatching {
-                    if (response.isSuccessful) {
-                        response
-                    } else {
-                        throw HttpException(response)
-                    }
-                })
-            }
+        emit(
+            suspendCancellableCoroutine { continuation ->
+                call.registerCallback(continuation) { response ->
+                    continuation.resumeWith(kotlin.runCatching {
+                        if (response.isSuccessful) {
+                            response
+                        } else {
+                            throw HttpException(response)
+                        }
+                    })
+                }
 
-            call.registerOnCancellation(continuation)
-        })
+                call.registerOnCancellation(continuation)
+            }
+        )
     }
 }