+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
-## Version 1.0.0 (2021-12.25)
+## Version 1.0.1 (2024-08-01)
 
-Initial release support kotlin 1.5.32 and coroutines 1.5.2
+Initial release use kotlin 1.9.24 and coroutines 1.8.1
+change min sdk 21.
+
+## Version 1.0.0 (2021-12-25)
+
+Initial release use kotlin 1.5.32 and coroutines 1.5.2